Day 049 - Room SceneNodes

+++ include/Dunjun/TickCounter.hpp
@@ -24,11 +24,11 @@ public:
 		return reset;
 	}
 
-	inline usize getTickRate() const { return m_tickRate; }
+	inline f64 getTickRate() const { return m_tickRate; }
 
 private:
 	usize m_tick = 0;
-	usize m_tickRate = 0;
+	f64 m_tickRate = 0;
 	Clock m_clock;
 };
